+++ src/ic-cdk/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Add `BufferedStableWriter` for efficient writing to stable memory (#245)
+
 ## [0.5.0] - 2022-03-29
 ### Added
 - Update canister calling API for 128-bit cycles (#228)